refactor existing remote key push out into keys/
