+++ wire/lib/src/hive/steps/mod.rs
@@ -1,4 +1,5 @@
 pub mod activate;
 pub mod build;
 pub mod evaluate;
+pub mod keys;
 pub mod push;
